change to google analytics v4 (#51)
* change to google analytics v4

* update sentry project name

* update sentry org
