+++ .env.template
@@ -12,6 +12,7 @@ REACT_APP_CODE_CHALLENGE_METHOD="S256"
 REACT_APP_HCAPTCHA_SITE_KEY="541838f2-e585-4726-b398-24102b1d4df8"
 REACT_APP_SENTRY_AUTH_TOKEN="fbd61f0af5c74fb88d27929e8c0975418d148773f17540ec8d663df19f85cf34"
 REACT_APP_SENTRY_DSN="https://791d7df9132d46a9868de0863bd6583c@o4504371258785792.ingest.sentry.io/4504371261079552"
+REACT_APP_GOOGLE_ANALYTICS_ID=""
 
 REACT_APP_SCROLL_ENVIRONMENT="STAGING"
 REACT_APP_CHAIN_ID_L1="5343511"
