+++ craco.config.js
@@ -60,8 +60,8 @@ module.exports = {
       })
       if (env === "production") {
         const sentryPlugin = new SentryWebpackPlugin({
-          org: "scroll-tech",
-          project: "scroll",
+          org: "scroll-zkp",
+          project: "scroll-frontends",
           include: "./build",
           authToken: process.env.REACT_APP_SENTRY_AUTH_TOKEN,
         })
