+++ package.json
@@ -39,6 +39,7 @@
     "react": "^18.2.0",
     "react-countdown": "^2.3.2",
     "react-dom": "^18.2.0",
+    "react-ga4": "^1.4.1",
     "react-helmet": "^6.1.0",
     "react-mailchimp-subscribe": "^2.1.3",
     "react-markdown": "^8.0.3",
@@ -163,7 +164,6 @@
     "qs": "^6.9.4",
     "react-device-detect": "^1.6.2",
     "react-feather": "^2.0.8",
-    "react-ga": "^2.5.7",
     "react-i18next": "^10.7.0",
     "react-popper": "^2.2.3",
     "react-redux": "^7.2.0",
