+++ src/pages/swap/components/SearchModal/CurrencySearchModal.tsx
@@ -1,5 +1,5 @@
 import React, { useCallback, useEffect, useState } from "react"
-import ReactGA from "react-ga"
+import ReactGA from "react-ga4"
 import { Currency } from "uniswap-v2-sdk-scroll"
 import useLast from "../../hooks/useLast"
 import { useSelectedListUrl } from "../../state/lists/hooks"
