+++ src/pages/swap/components/SearchModal/ListSelect.tsx
@@ -1,6 +1,6 @@
 import React, { memo, useCallback, useMemo, useRef, useState } from "react"
 import { ArrowLeft } from "react-feather"
-import ReactGA from "react-ga"
+import ReactGA from "react-ga4"
 import { usePopper } from "react-popper"
 import { useDispatch, useSelector } from "react-redux"
 import { Text } from "rebass"
