+++ src/pages/swap/index.tsx
@@ -1,7 +1,5 @@
 import "inter-ui"
-import React, { StrictMode } from "react"
-import { isMobile } from "react-device-detect"
-import ReactGA from "react-ga"
+import { StrictMode } from "react"
 import { Provider } from "react-redux"
 import "./i18n"
 import App from "./pages/App"
@@ -16,23 +14,6 @@ import ThemeProvider, { FixedGlobalStyle, ThemedGlobalStyle } from "./theme"
 if ("ethereum" in window) {
   ;(window.ethereum as any).autoRefreshOnNetworkChange = false
 }
-
-const GOOGLE_ANALYTICS_ID: string | undefined = process.env.REACT_APP_GOOGLE_ANALYTICS_ID
-if (typeof GOOGLE_ANALYTICS_ID === "string") {
-  ReactGA.initialize(GOOGLE_ANALYTICS_ID)
-  ReactGA.set({
-    customBrowserType: !isMobile ? "desktop" : "web3" in window || "ethereum" in window ? "mobileWeb3" : "mobileRegular",
-  })
-} else {
-  ReactGA.initialize("test", { testMode: true, debug: true })
-}
-
-window.addEventListener("error", error => {
-  ReactGA.exception({
-    description: `${error.message} @ ${error.filename}:${error.lineno}:${error.colno}`,
-    fatal: true,
-  })
-})
 
 function Updaters() {
   return (
