+++ src/pages/swap/pages/AddLiquidity/index.tsx
@@ -2,7 +2,7 @@ import { BigNumber } from "@ethersproject/bignumber"
 import { TransactionResponse } from "@ethersproject/providers"
 import React, { useCallback, useContext, useState } from "react"
 import { Plus } from "react-feather"
-import ReactGA from "react-ga"
+import ReactGA from "react-ga4"
 import { RouteComponentProps } from "react-router-dom-v5"
 import { Text } from "rebass"
 import { ThemeContext } from "styled-components"
