+++ src/pages/swap/pages/MigrateV1/MigrateV1Exchange.tsx
@@ -1,7 +1,7 @@
 import { TransactionResponse } from "@ethersproject/abstract-provider"
 import { AddressZero } from "@ethersproject/constants"
 import React, { useCallback, useMemo, useState } from "react"
-import ReactGA from "react-ga"
+import ReactGA from "react-ga4"
 import { Redirect, RouteComponentProps } from "react-router"
 import { Text } from "rebass"
 import { Currency, CurrencyAmount, Fraction, JSBI, Percent, Token, TokenAmount, WETH } from "uniswap-v2-sdk-scroll"
