+++ src/pages/swap/pages/MigrateV1/RemoveV1Exchange.tsx
@@ -3,7 +3,7 @@ import { TransactionResponse } from "@ethersproject/abstract-provider"
 import { AddressZero } from "@ethersproject/constants"
 import { Contract } from "@ethersproject/contracts"
 import React, { useCallback, useMemo, useState } from "react"
-import ReactGA from "react-ga"
+import ReactGA from "react-ga4"
 import { Redirect, RouteComponentProps } from "react-router"
 import { CurrencyAmount, Fraction, JSBI, Percent, Token, TokenAmount, WETH } from "uniswap-v2-sdk-scroll"
 import { ButtonConfirmed } from "../../components/Button"
