+++ src/pages/swap/pages/RemoveLiquidity/index.tsx
@@ -3,7 +3,7 @@ import { Contract } from "@ethersproject/contracts"
 import { TransactionResponse } from "@ethersproject/providers"
 import React, { useCallback, useContext, useMemo, useState } from "react"
 import { ArrowDown, Plus } from "react-feather"
-import ReactGA from "react-ga"
+import ReactGA from "react-ga4"
 import { RouteComponentProps } from "react-router"
 import { Text } from "rebass"
 import { ThemeContext } from "styled-components"
