+++ src/pages/swap/pages/Swap/index.tsx
@@ -1,6 +1,6 @@
 import React, { useCallback, useContext, useEffect, useMemo, useState } from "react"
 import { ArrowDown } from "react-feather"
-import ReactGA from "react-ga"
+import ReactGA from "react-ga4"
 import { Text } from "rebass"
 import { ThemeContext } from "styled-components"
 import { CurrencyAmount, JSBI, Token, Trade } from "uniswap-v2-sdk-scroll"
